+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.21.5...HEAD) - TBD
+## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.21.6...HEAD) - TBD
+
+## [4.21.6](https://github.com/schemathesis/schemathesis/compare/v4.21.5...v4.21.6) - 2026-06-14
 
 ### :bug: Fixed